chore: reorganize module
